Añadido modelo de Comentario, TransicionTicket y Adjunto

+++ backend/src/main/java/com/arelance/helpdesk/modelo/Comentario.java
@@ -29,7 +29,7 @@ public class Comentario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idComentario")
-    private Integer idComentario;
+    private Long idComentario;
 
     @Column(nullable = false)
     private String texto;
